requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-dev -o requirements.txt
aiodns==4.0.0
    # via auto-vpngate
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.5
    # via auto-vpngate
aiosignal==1.4.0
    # via aiohttp
attrs==26.1.0
    # via aiohttp
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via
    #   cryptography
    #   pycares
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
cryptography==46.0.6
    # via google-auth
cssselect==1.4.0
    # via pyquery
flask==3.1.3
    # via auto-vpngate
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.30.2
    # via
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
google-cloud-core==2.5.1
    # via google-cloud-firestore
google-cloud-firestore==2.26.0
    # via auto-vpngate
googleapis-common-protos==1.74.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.80.0
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   grpcio-status
grpcio-status==1.80.0
    # via google-api-core
gunicorn==25.3.0
    # via auto-vpngate
idna==3.11
    # via
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
lxml==6.0.2
    # via pyquery
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
packaging==26.0
    # via gunicorn
prettytable==3.17.0
    # via auto-vpngate
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.2
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.33.6
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycares==5.0.1
    # via aiodns
pycparser==3.0 ; implementation_name != 'PyPy'
    # via cffi
pyquery==2.0.1
    # via auto-vpngate
requests==2.33.1
    # via google-api-core
typing-extensions==4.15.0
    # via grpcio
urllib3==2.6.3
    # via requests
wcwidth==0.6.0
    # via prettytable
werkzeug==3.1.8
    # via flask
yarl==1.23.0
    # via aiohttp
